refactor: 게시물 삭제 시 존재하는 게시물인지 체크하는 로직 추가

+++ src/main/java/com/blog/be/post/application/PostCommandService.java
@@ -102,7 +102,11 @@ public class PostCommandService {
     public void deletePost(Long postId) {
         postTagRepository.deleteAllByPostId(postId);
 
-        String postUuid= postRepository.deleteById(postId);
+        if (!postRepository.existsById(postId)) {
+            throw new PostException(PostErrorCode.POST_NOT_FOUND);
+        }
+
+        String postUuid = postRepository.deleteById(postId);
 
         applicationEventPublisher.publishEvent(new PostDeletedEvent(postUuid));
     }
